nice fix for Issue #5
This fix evaluates all should statements in a chain and only throws at the end if one of the shoulds failed. In order to provide a stacktrace with the failing should I introduced a custom type testresult. Testresult's Status (Passed | Failed) is also returned when using the -test switch (instead of True | False previously). The testresult is sent through the pipe along with the input and stripped from the input before the original code evaluates the input.